[FIX] hr_appraisal: fix failing tests
Since this [commit] the employee work_email and phone are stored in a separate
res.partner record. When a user is linked to an employee, the partner related
to both the employee and the user should be the same. When appraisal requests
are created in the appraisal tests, the tests assume the name of the recipients
equal those of the corresponding employees. However, a recipient is a partner,
not an employee. More specifically it is the partner linked to the employee and
to the associated user. As such the same names should be used when creating the
users and their corresponding employees.

opw-3031187

[commit]: https://github.com/odoo/odoo/commit/3c6060b7bbe9c67aca8073ef43c1e89fb7e820ca

closes odoo/enterprise#33565

Related: odoo/odoo#103792
Signed-off-by: Kevin Baptiste <kba@odoo.com>