[FIX] social_linkedin: use the versioned API
Purpose
=======
Unversioned API will be depreciated and we need to specify the version
of the API we want to use.

Note that the versioned API is supported for 12 months, and so we might
need to re-update the version number...

Change the way we upload images, create post and fetch posts.

Task-3054273

closes odoo/enterprise#35492

X-original-commit: 0fb3bb9ec6699bd84593aa615e373d144a4cb58d
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>