[FIX] social_linkedin: use the versioned API
Purpose
=======
Unversioned API will be depreciated and we need to specify the version
of the API we want to use.

Note that the versioned API is supported for 12 months, and so we might
need to re-update the version number...

Change the way we upload images, create post and fetch posts.

Task-3054273

closes odoo/enterprise#35514

X-original-commit: 152a8df8b257e68c3e7fd79c9b84a9a9e7199ac2
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>