[REM] event_sale_dashboard: remove module
All fields and the override of `read_group` of `EventSaleReport`
is not used since 70412a6f908d04985d93ec21f5031c341d3eabc8.
And the only remaining view is inherited with 'primary' mode,
it is not the default view of the model and it is unused.
Then remove all the module.

closes odoo/enterprise#33568

Related: odoo/upgrade#4012
Related: odoo/odoo#104863
Signed-off-by: Stéphane Debauche (std) <std@odoo.com>