[FIX] mrp_workorder(plm): fix worksheet document
In the 'Add a Step' operation, the uploaded document is not correctly
displayed in the relevant views.

Not all required data items were transmitted and the 'Tablet' component
did not fully handle those data items, especially with the PLM module
installed.
Since handling images would have required new fields, this
functionnality will require a larger refactor and is deferred to master,
which is why a filter on pdf files has been introduced.

closes odoo/enterprise#33700

Task: 2985735
X-original-commit: 8c8881115fb2666ea9144938a360e3b3b8344022
Signed-off-by: William Henrotin (whe) <whe@odoo.com>