[REF] various: cleanup usage of mail.composer.mixin
PURPOSE

Purpose of this commit is to cleanup usage of mail.composer.mixin, remove
unnecessary stuff and use its features when possible.

ACCOUNT FOLLOWUP

Remove unnecessary domain force on template_id field. It is done using
base render_model field available in mixin and correctly overridden in
this addon.

Add a domain on sms template, as this one does not benefit from any mixin
currently.

Finally remove unused 'email_add_signature' field, as it is not used.
Indeed either a template is used and it should contain the signature,
either a default body is given in '_get_main_body' which already contains
a signature.

HR_APPRAISAL(_SURVEY)

Email_from fields are not necessary as anyway we always use the current
user's email and field is not present in form view. Override of default_get
to raise about invalid email_from is not necessary as anyway it raises
when posting a message. Giving author_id to message_post is sufficient
as it computes email_from based on author when given.

Add render_model in views, as it is part of the domain given to template_id
field of composer mixin.

Use _render_field when possible instead of _render_template as its API
is simpler and uses field definition for rendering parameters.

Task-2710804 (Mail: Clean MailThread Posting API)

Part-of: odoo/enterprise#31017