[FIX] website_knowledge: use fields of the articles during search
Since [1] a non-existing `content` field was used to populate the `body`
information of search results about `knowledge.article`.
It made the search snippet crash (if set to search in everything) and
impossible to be used anymore if you had website_knowledge installed.

Also, its URL was not fetched, making the search result lead to the
current page.

This commit gets the body data from the `body` field and fetches the
`website_url` field so that the search result links can be clicked on.

[1]: https://github.com/odoo/enterprise/commit/18b643e72dc5713f79f501b0d972452b07457f49

Related: https://github.com/odoo/odoo/pull/104962

opw-3045092

Part-of: odoo/enterprise#33585