[FIX] helpdesk: fix searching on sla_success field
Before this commit:

In the case where a ticket contains both a success and a failed SLA,
that ticket is show in both SLA Success and SLA Failed filter.

After this commit:

A ticket will be considered failed if it contains at least one failed
SLA and it'll not be show in SLA Success filter.

task-3000817

X-original-commit: 5f2d57886bfcfaf4f5b445477cd2a901a3ff55bc
Part-of: odoo/enterprise#36262