[FIX] helpdesk: fix searching on sla_success field
Before this commit:

In the case where a ticket contains both a success and a failed SLA,
that ticket is show in both SLA Success and SLA Failed filter.

After this commit:

A ticket will be considered failed if it contains at least one failed
SLA and it'll not be show in SLA Success filter.

task-3000817

X-original-commit: https://github.com/odoo/enterprise/pull/33555/commits/67f3155078d6050f874dde41b87a93bd48583e4c
Part-of: odoo/enterprise#36222