[FIX] sale_subscription: recurring monthly was not recomputed at renewal start
Before this commit, the compute depens of _compute_recurring_monthly and
_compute_recurring_live were not correct.

When a contract is renewed, it's start_date is in the future.
Before this commit, the MRR of the renewal is expected to be 0 until the start_date is passed.
Then it is computed according to the lines MRR.
It is confusing and it did not worked. Now the MRR is not null even if the start_date is not
passed but the recurring_live value is updated according to the parent state and stage_category.
Relying on the parent state allows to avoid using a daily cron to check if renewal are becoming alive
according to their state and start_date.

When the renewed (parent) contrat is closed because it reached the end_date, the renewal will see
his recurring_live value be recomputed and it will be True.

This commit also update the sale.order.log behavior to keep the current behavior and the code is slightly
simplified.

X-original-commit: df01f007fd5c282462ea58ddf3e1b92837fdb2ae
Part-of: odoo/enterprise#34616