[FIX] account_accountant: fixing view and dark mode
The view was broken since the recent css change, the notebook manual operations is composed by two tables and one was below the other.

Thanks to this PR it's now fixed, the two tables are now next to each other.

Moreover the progress bar was just a purple square, the behaviour of v15.0 is now restore + adapted to dark mode

This PR also correct multiple dark mode problem
* White rectangle
* Border
* Hover Line
* Hover Field

Task-id: 3055687
Part-of: odoo/enterprise#33699