[IMP] stock_barcode*: remove added props
Into the `MainComponent`'s props, there was two values added in the
`setup`: `model` and `id`.
It isn't a good practice to add props like that, all the props should be
passed by the parent only.
Instead, there are replaced by usual `Class` properties, and are renamed
`resId` and `resModel` to be less ambiguous.

task-3048127

Part-of: odoo/enterprise#33600