[IMP] mrp_plm: Adapt method signatures
Since fixes done on the bom overview had to be done in a released
version, no method signature changes could be donce, hence the use of
excessive context keys to bypass it.

Changes the signatures so that future work is done on better grounds,
and so that each method explicitely asks for what parameters it needs to
do its work.

closes odoo/enterprise#34577

Related: odoo/odoo#107006
Signed-off-by: Arnold Moyaux (arm) <arm@odoo.com>