[FIX] l10n_es_reports: fix balance sheet unaffected earnings
Currently, there is a miscalculation of the unaffected earnings.
In particular with account '129000 Resultado del ejercicio',
one could post a journal entry in the previous fiscal year.
And that was not handeled correctly and lead to inaccurate reports.

This PR is the combination of internal bug reports, and this ticket:
opw-2950285

closes odoo/enterprise#33756

X-original-commit: 7fd2906135690b68734d4f861a8185ec9f27f9e0
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>
Signed-off-by: John Laterre (jol) <jol@odoo.com>