[FIX] l10n_es_reports: fix balance sheet unaffected earnings
Currently, there is a miscalculation of the unaffected earnings.
In particular with account '129000 Resultado del ejercicio',
one could post a journal entry in the previous fiscal year.
And that was not handeled correctly and lead to inaccurate reports.

This PR is the combination of internal bug reports, and this ticket:
opw-2950285

closes odoo/enterprise#33644

Signed-off-by: Olivier Colson (oco) <oco@odoo.com>