[FIX] account_invoice_extract: don't recreate the boxes on reload
Previously, the boxes were re-created upon clicking on the "Reload
AI Data" button.

This caused a traceback when clicking on a box in the interface as
it wasn't aware of the new boxes and the ID linked to it belonged
to a record that has been unlinked.

Now, the boxes are kept so that the IDs are still the same.
This has the downside that there might be a mismatch between was has
been fetched during the reload and the old boxes, but it is very
unlikely to occur.

closes odoo/enterprise#33804

X-original-commit: 6f1fecfc034c2f98e240028a7c158d0c4e4934d2
Signed-off-by: Florian Daloze (fda) <fda@odoo.com>
Signed-off-by: Louis Baudoux (lba) <lba@odoo.com>