[FIX] studio: fix web_editor inline error
Inserting an inline block next to a regular block in Studio's report editor and
then editing the block's content is does not work as expected (the text
ceases to be inline).

This happens because web_editor wraps the content of this inline block in <p>
tags.

This behavior was introduced by a fix that wraps any inline nodes into a p
(paragraph tag) at the root of the editable (see PR odoo/odoo #93207), and
it's the Editor's default behavior.

In order to disable this behavior in Studio's report editor, a new option
was added to OdooEditor's options object.

task #2995601

closes odoo/enterprise#35013

X-original-commit: 680c2e21310fa074d67842eb73eb2e80b0b75ed4
Related: odoo/odoo#108068
Signed-off-by: David Monjoie (dmo) <dmo@odoo.com>