[FIX] account_accountant: removing duplicate function + typo + dark mode fix
The view was broken since the recent css change, the notebook manual operations is composed by two tables and one was below the other.

Thanks to this PR it's now fixed, the two tables are now next to each other.

Moreover the progress bar was just a purple square, the behaviour of v15.0 is now restore + adapted to dark mode

This PR also correct multiple dark mode problem

White rectangle
Border
Hover Line
Hover Field
The second commit is about duplicate function that needed to be removed.

closes odoo/enterprise#33937

Task-id: 3055687
Forward-port-of: #33699
Signed-off-by: Florian Gilbert (flg) <flg@odoo.com>
Signed-off-by: Maximilien La Barre (malb) <malb@odoo.com>