[FIX] knowledge: do not use toolbar title as default title
When an article does not have any title and the user clicks on the title,
the script will retrieve the first heading of the editor and use it as
default title. When the user uses a mobile device, wysiwyg will add a
toolbar in the editor and the selector used to fetch the first heading
of the editor will match with a title of that toolbar. To avoid that, we
will update the selector to only select a heading from the editable zone.

task-3092930

closes odoo/enterprise#34892

X-original-commit: ce4d1d6c2d5e21b385fbf0e147df01738263af7d
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>