[FIX] mrp_workorder: remove unnecessary, forbidden directive
Form view archs should not contain `t-att-` directives, as it is
not qweb. Worse, they can't use implementation details of the js
view used to render it (in this case `isBlocked`).

Since `isBlocked` isn't defined in the rendering context anyway,
the button was never disabled via this directive, so removing
it has no effect. Functionnaly speaking, it works though, because
the button is actually disabled in javascript.

X-original-commit: 31d76286a11798aae916dcb2f786bc9d1388d09f
Part-of: odoo/enterprise#33841