[FIX] knowledge: disable form arch warnings
This commit comes with odoo/odoo#103722

The knowledge main action uses a form view. In its arch, a lot of
implementation details are used (owl directives/components, functions
or attributes of the renderer). That should never be used in archs.
PR odoo/odoo#103722 introduces warnings to detect and prevent people
from doing that. However, the change would be to large for a stable
to fix knowledge. For that reason, this commit disables the warnings.
Knowledge should be reworked in master.

Part-of: odoo/enterprise#35065