[FIX] delivery_easypost: unnecessary raise
Whenever we create a sales order with the shipping configured with EasyPost it estimates the price.
Imagine we create a sales order, confirm it and then we realise the made a mistake with one of the products. Given that the picking was already confirmed, we cannot delete the line, however, we can change its quantity to 0.
If we change the quantity of one of the products to 0, we currently get an error "The estimated price cannot be computed because the weight of your product %s is missing.".
Such raise is not needed in this particular case.
If all the quantities are 0, there is already another error that will be raised before. If some of the products are 0, the weight is also 0 and it will not impact the package.
The error will still be raised if the quantity is not 0 and the weight has not been defined on the product.

closes odoo/enterprise#34732

X-original-commit: 9c0810da7daabb99b1896d88a922ebdbc46a45e5
Signed-off-by: Simon Goffin <sig@odoo.com>