[FIX] industry_fsm_report: worksheet templates for tasks duplication
Steps to reproduce:

- We have to go to any project with subt-tasks and workseets enabled.
- We create a new task with a few sub-tasks inside and select manually
a worksheet template for each sub-task.
- We duplicate this task.

Issue:

When we duplicate the task we lose the sub-tasks worksheet templates and
they're replaced by the default project worksheet template.

Solution:

Following the same flow of computing the worksheets templates in project
, we have to add the proper condition to not lose the worksheet template
of the task if it has one already.

This bug affects all versions from saas-15.2 until 16.

opw-3058207

closes odoo/enterprise#33746

Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>