[FIX] l10n_lu_reports: Fix Balance Sheet(s)
Luxembourgish balance sheets (both abridged and full) differ between versions of Odoo.

This PR correct those report and to be more precise, it corrects two Report Lines ("the Profit or loss brought forward" and "Profit or loss for the financial year").

Removing the possibility to fold/unfold those line + change in the formula.

closes odoo/enterprise#39130

Task-id: 2973476
Signed-off-by: William André (wan) <wan@odoo.com>