[FIX] mrp_workorder_hr(_account): AA line for employee cost incorrect
Previously we only post analytic accounting entries for employee cost
when close a mrp.workcenter.productivity log. This makes the AA lines
are not correctly posted when we manually create a log or change the
time on an existing log.

In this commit, we change it to post aa lines when computing duration
for mrp.workcenter.productivity log.

We also make employee_cost on mrp.workcenter.productivity computed
instead of assigning it when closing the log, so that we always have the
value when we calculate the cost for aa lines.

X-original-commit: 1ac42e9a38da1fa0e5013912690d8203dc1c6e80
Part-of: odoo/enterprise#34098