[FIX] account_reports: show all lines on report form
This is necessary for the resequence widget to work properly. Without that, since the report lines have no sequence defined in the data files (and are therefore ordered by id by default), the sequence widget only assigns sequences to the current page of the pager, hence leading to inconsistent and buggy sequencing of the lines.

The sequence widget cannot be fixed in stable (too complex according to dbo and fge), so we patch this by setting a limit of 2000 lines per page (assuming no report will go that high anyway).

closes odoo/enterprise#33780

X-original-commit: 9f3a471c55d56a0b8da0366b508421384c0b6082
Signed-off-by: Laurent Smet <las@odoo.com>
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>