[FIX] website_helpdesk: enable to create a ticket when submitting forms
This commit makes the option "Create a ticket" work when creating forms.

Steps to reproduce the bug:
- Create a form where an user can fill in some info for customer support
  on the website (select "Create a ticket" as action so that once
  submitted, it creates a ticket in the helpdesk app automatically).
- The form actually disappears.

In order to work properly, the JS module `website_helpdesk.form`
requires the JS module `website.form_editor_registry`. That one is
defined in its manifest inside the asset called `website.assets_editor`.
The module `website_helpdesk.form` should thus be defined in
`website.assets_editor` as well.

This was the case in 15.0 but then [1] merged the website_helpdesk_form
module into website_helpdesk and made the mistake of moving the related
file inside the wrong asset (assets_frontend). This however worked
because both assets_frontend and assets_editor are loaded when using the
editor. The only side-effect was then that visitors loaded that editor
file for no reason (and it fails silently). This is however not the case
anymore since [2] when the website and editor UI were moved inside the
backend: the website iframe uses the assets_frontend bundle, while the
backend/editor uses assets_editor. This could be fixed in 15.X but as
explained, the only real bug occurs from 16.0 so this fixes the error in
that version only.

[1]: https://github.com/odoo/enterprise/commit/d26c98bdc2dbe7541a12c5bdf07baf889165b661
[2]: https://github.com/odoo/odoo/commit/31cc10b91dc7762e23b4bde9b945be0c4ce3fe3b

opw-3054728

closes odoo/enterprise#33764

Signed-off-by: Quentin Smetz (qsm) <qsm@odoo.com>