[IMP] stock_barcode: scan location to add new stock move line
This commit improves handling of reserved quantities in the barcode app,
it allows the user to split move lines to reserve product from different
locations than the one specified, this is done by scanning a location after scanning
some products. The missing quantity from the original line will be reserved
by the new line and are to be scanned on the new line.
The original line reserved quantity will be set
to the qty done and marked completed.

Adds a test to check the quantites reserved by the split move lines.

closes odoo/enterprise#34652

Taskid: 2859690
Signed-off-by: Steve Van Essche <svs@odoo.com>