[IMP] stock_barcode: Main Component use state view
Before this commit, the variable used to define what should be shown,
`view`, was set on the BarcodeModel, and every time it was modified, the
model triggered an event 'update' to re-render the MainComponent.
It seems better (and more "OWLish") for `view` to be a property of the
MainComponent, and that it uses the `state` to be re-rendered every time
this property is modified instead of manually calling `render`.

Same for the variable showing the picking's note, it's simpler to use
the state.

task-3048127

Part-of: odoo/enterprise#33600