[FIX] l10n_uk_reports: Add missing fields in fraud detection header
When submitting the tax report in the UK, some information should be
provided in the header of the request. Some of them where either
missing or wrongly formatted.

A new version of the send button was required in order to get client
side information that are required (such as the size of the screen and
the size of the window).

Note that errors about the header for the multi-factor authentification
and about the license can be ignored because it's not mandatory to
activate such security when using Odoo and not everyone using Odoo has
a enterprise license (discuss with HMRC directly)

The list of headers and their format can be found here:
https://developer.service.hmrc.gov.uk/guides/fraud-prevention/connection-method/web-app-via-server/

How to Test:
  1. Activate Debug Mode (in hmrc_service.py)
  2. In hmrc_vat_obligation.py, the endpoint request needs to be changed to /test/fraud-prevention-headers/validate
  3. An test individual needs to be created using the government page: https://developer.service.hmrc.gov.uk/api-test-user
  4. The VAT Registration Number needs to be copied into the UK company. Keep the username and password too
  5. When submitting the tax report, you'll first need to connect to hmrc using the credentials previously gotten.
NB: The credentials can be reset by going to the user page and under the UK HMRC Integration tab.

Known issues:
  - When testing in local, some of the headers won't be valid (for example
  the gov-client-public-ip header can't be provided, gov-vendor-forwarded won't
  be a public IP address... Every header that require Public IP address or
  different vendor and client IP address)
  - Multifactor header can't always be provided because the option is
  not mandatory to use Odoo
  - The Gov-Vendor-License-IDs also can't always be provided

task-2925880

closes odoo/enterprise#34247

Signed-off-by: Brice Bartoletti (bib) <bib@odoo.com>