[IMP] l10n_be_hr_contract_salary: Show 'new_car' only when needed
Purpose:
Generally, the user is not able to pick the new car if the amount of used
cars is already reached. But it can be overruled by ticking the field 'new_car'
on the 'Generate a Simulation Link' wizard. Thus it makes sense to show
'new_car' only when it is needed.

Now 'new_car' is always visible on the wizard.

After this commit the field 'new_car' will only be visible on the wizard if
the amount of used cars is already reached.

task - 3035820

closes odoo/enterprise#33788

Signed-off-by: Kevin Baptiste <kba@odoo.com>