[ADD] l10n_pe_edi_stock_20: Peru delivery guide 2.0
SUNAT Resolution 000123-2022, which came into force on 1 December 2022,
provides a new framework for the filing of 'Guías de Remision'
('delivery guides') by companies.

Any company that transfers goods (e.g. for a sale or from one location
to another) must file a 'Guía Remitente'.
In addition, there is a specialized regime for carriers (companies
specialized in the transport of goods) which must file a
'Guía Transportista'.

After filing, SUNAT provides a confirmation that includes a QR code that
must be presented by the driver if stopped by the authorities.

For creating delivery guides, SUNAT provides an online portal and a
REST API. The REST API has three endpoints: one for requesting an
authentication token, one for filing a delivery guide, and one for
retrieving the confirmation.

This module implements the filing of 'Guía Remitente' via the REST API,
and the retrieval of the confirmation including the QR code.

Implementation notes:
- We reused the existing module l10n_pe_edi_stock as much as possible.
  The old module is no longer in use since the old API was deprecated.
  We plan to merge the two in master.
- Often, the confirmation is not yet ready when we attempt to retrieve
  it. We catch that error code; in that case, the 'Retry' button
  attempts to re-retrieve the confirmation.
- If the filing/confirmation process fails, the 'Retry' button will
  file a new delivery guide with a new document number.
- Several new fields have been added, including:
  - credentials (the credentials for the REST API are different from
    those for the invoice EDI);
  - ID / authorization number fields for the company, the vehicle and
    the transport operator;
  - 'tariff fraction code' field on products, which needs to be
    specified if the product is a normalized product; and
  - Related Document type/number fields on the picking, to reference a
    supporting document for the transfer (e.g. invoice or authorization)

closes odoo/enterprise#35228

Task: 3068389
X-original-commit: c7001ceb254afca59dba4bd0e79e6d6783b8a5b3
Signed-off-by: Josse Colpaert <jco@odoo.com>