[FIX] web_studio: a WOWL editor for the new ListView
Before this commit, the editor for the listView was still in legacy,
giving way to some inconsistencies or crashes, when field widgets
and widgets were present in the new WOWL framework, but not in legacy.

After this commit, a new editor is introduced that fully uses all features provided
by the WOWL framework.

It is built around the same premises as the formEditor[1] and the kanbanEditor[2].
Those includes the adjunction of a xpath for each relevant aspect of the listView (fields, widgets),
in order to be able to edit those in the arch, and ease the the editing experience in general.

[1] 370ba732e84b8d121e3283a16dbf233d0c7ac4c8
[2] 3229150af252ba40a8808bc95d74f09503f2d841

closes odoo/enterprise#34074

X-original-commit: 53c59e79d99645ce83cbe34d2d1fdb3142088f7b
Related: odoo/odoo#105970
Signed-off-by: Aaron Bohy (aab) <aab@odoo.com>
Signed-off-by: Lucas Perais (lpe) <lpe@odoo.com>