[FIX] l10n_uk_reports: fix _custom_options_initializer()
Untill now, in order to have the HMRC button visible on
the tax report, we had to set the `availability_condition` to `always`.
Which does not make much sense.

So this condition is removed here, along with a check on the
`account_fiscal_country_id.code` that has become useless since
`_custom_options_initializer()` is now located in the
`l10n_uk.tax.report.handler`.

closes odoo/enterprise#33816

X-original-commit: f3125089c74cd85536b99d8e419c7adb871a2110
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>
Signed-off-by: John Laterre (jol) <jol@odoo.com>