[IMP] social(_*): convert t-on-click from views arch
Kanban view archs should not contain `t-on-click` directives as
it is not a qweb directive.
Also, it's forbidden to use JS implementation details that are
only available in the custom js_class the main reason for that
is that it will crash if used by studio because js_class are
ignored by it.

This commit converts all `t-on-click` occurences in the views
arch by addEventListener in the KanbanRenderer for the Image
Carousel and in the KanbanRecord for the Comments / Likes
feature of the various social medias.

Task-3061215

Part-of: odoo/enterprise#34228