[FIX] documents: multiple selection on mobile doesn't require touchevent
This commit adapts the Documents ListView mobile's test to the related
community PR odoo/odoo#99318.

In a nutshell, triggering a "click" should properly simulate the actual
selection behavior (like in other asserts), so touch{start,end} event
aren't required.

Also, as the `click()` helper already waits for a `nextTick`, there is
no reason to wait for a second one.

closes odoo/enterprise#33839

Signed-off-by: Pierre Paridans (app) <app@odoo.com>