[IMP] sale_subscription: use sale's generic _create_invoices
The goal of this PR is to de-cluter the massive subscription
invoicing function.

To achieve this, this commit change the way we manually invoice
subcription.
Instead of using the `_create_recurring_invoice` we now use
sale's generic `_create_invoices` button for sale_subscription.

This commit also clean the _create_recurring_invoice at the same
time to avoid having test specific code as much as possible.

task-id : 3058418

Part-of: odoo/enterprise#34017