[FIX] sale_stock_renting: make RentalOrderLine.write overridable
before this commit super was called in the middel of the write function,
making it impossible to properly override the function. The only methods
to everride the function would be rewriting it completely and tempering with the MRO or calling the
parent-class of RentalOrderLine dyrectly, which are methods that should
be avoided.

after this commit, the portion after the super call is isolated into a
seperate function, making it possible to override the portion before and
after super seperately.

closes odoo/enterprise#34072

X-original-commit: cfe7cb7fc914c0193baeb653385f576c97e43cec
Signed-off-by: Victor Feyens (vfe) <vfe@odoo.com>