[FIX] website_sale_subscription: subscription info in qweb context
Before this commit, the Qweb context in ecommerce had no information
about the sale order being a subscription or not. This causes UI
problem, such as not the displaying the message warning that payment
details are saved automatically when paying an active subscription.

After this commit, the Qweb context dictionary contains a new key
`is_subscription` in order to know whether the sale order is a
subscription or not.

closes odoo/enterprise#33840

Signed-off-by: William Braeckman (wbr) <wbr@odoo.com>