[FIX] website_sale_subscription: subscription info in qweb context
Before this commit, the Qweb context in ecommerce had no information
about the sale order being a subscription or not. This causes UI
problem, such as not the displaying the message warning that payment
details are saved automatically when paying an active subscription.

After this commit, the Qweb context dictionary contains a new key
`is_subscription` in order to know whether the sale order is a
subscription or not.

closes odoo/enterprise#33871

X-original-commit: 0c7552267314531ea41e9f1f030d354654e83838
Signed-off-by: William Braeckman (wbr) <wbr@odoo.com>
Signed-off-by: Vallaeys Valentin (vava) <vava@odoo.com>