[FIX] account_report, account_accountant: Fix breadcrumbs manual reco
The breadcrumbs of the manual reconciliation widget was broken before this PR. It was displayed "unnamed" which is not clear for the user.

This PR correct that by putting "Reconciliation" instead of "unnamed" which is the behaviour of the flow Accounting -> Reconciliation.

The problem is that the name is not store in the database so the breadcrumbs become "unnamed".
To deal with that, I've added the name to the function that update the control panel and breadcrumbs and the init function. I've also written a getTitle function that will be used to update the "controlPanelProps.title".

closes odoo/enterprise#34052

X-original-commit: 1673e9a7ae7e1f9cd4ed4709633159aba0fe620f
Signed-off-by: Florian Gilbert (flg) <flg@odoo.com>
Signed-off-by: Maximilien La Barre (malb) <malb@odoo.com>