[FIX] stock_barcode: scanning product tracked by lot mistmatch quants without lot
Current behavior:
When you scan a product in the Barcode app that is tracked by lots,
it will match quants that does not have lot_id. But when you scan
the lot barcode, it will count as the same product as the previous one,
and it will try to change the lot_id of the quant, which is not
possible.

Steps to reproduce:
  1. Install Barcode and Inventory app
  2. Enable tracking by lot in the settings
  3. Create a new product, give it a barcode and track it by lot
  4. Add quantity on hands without any lot id
  5. Create a new lot id
  6. Go to the barcode app and do an inventory adjustment
  7. Scan the product barcode
  8. Scan the lot barcode
  9. Confirm the inventory adjustment (JS error)

After this commit:
When scanning the lot id, it will prevent matching lines having having
the same product but no lot id. This way, the line won't end up editing
the lot of an existing quant.

opw-3023190

closes odoo/enterprise#35633

X-original-commit: 77aaa523900ecb9fbf9bbf39c4c245248a38c629
Signed-off-by: Steve Van Essche <svs@odoo.com>
Co-authored-by: Steve Van Essche <42638301+svs-odoo@users.noreply.github.com>