[FIX] l10n_co_edi: prevent long strings of text by breaking components
Before this commit
==================
When an invoice is sent for electronic invoice validation and the EDI State is
changed to Sent, the view displays the field `CUFE/CUDE(l10n_co_edi_cufe_cude_ref)`,
the value related to this field is a long string and it's affecting the view.

After this commit
=================
I have added the class in the `CUFE/CUDE(l10n_co_edi_cufe_cude_ref)` field and
prevented long strings of text by breaking components.

TaskID - 3053822

closes odoo/enterprise#33928

X-original-commit: ed163a2f9e46c6897d3367f4c13cd2181fed045e
Signed-off-by: Josse Colpaert <jco@odoo.com>
Signed-off-by: Bhavin Patel (bat) <bat@odoo.com>