[FIX] l10n_co_edi: prevent long strings of text by breaking components
Before this commit
==================
When an invoice is sent for electronic invoice validation and the EDI State is
changed to Sent, the view displays the field `CUFE/CUDE(l10n_co_edi_cufe_cude_ref)`,
the value related to this field is a long string and it's affecting the view.

After this commit
=================
I have added the class in the `CUFE/CUDE(l10n_co_edi_cufe_cude_ref)` field and
prevented long strings of text by breaking components.

TaskID - 3053822

closes odoo/enterprise#33890

Signed-off-by: Josse Colpaert <jco@odoo.com>