[IMP] social(_*): remove _insertThousandSeparator method
the basic integer field already adds thousands separator so we
don't need another method to do it. And it was calling a js_class
method which is forbidden for the sake of studio.
This commit converts the <b> tags with a function call to a
simple <field> tag.

Task-3061215

Part-of: odoo/enterprise#33806