[IMP] social(_*): remove _insertThousandSeparator method
the basic integer field already adds thousands separator so we
don't need another method to do it. And it was calling a js_class
method which is forbidden for the sake of studio.
This commit converts the <b> tags with a function call to a
simple <field> tag.

Task-3061215

X-original-commit: 03b200e912c9b7f7d9ec139ea75b55c7bac847b7
Part-of: odoo/enterprise#34687