[FIX] mrp_workorder_plm: Allow basic users to make step suggestions
In the tablet view, a user can make steps suggestions (such as creating
a new step, deleting a step, ...). This is supposed to work for basic
users (not only admins), so workers can make feedback directly in the
workorder as they are doing them.

When PLM is added, it also adds an override so that instead of writing
an activity on the BOM, it directly creates an ECO with the suggested
step modification inside.

The issue is that :
- It requires "users" access levels to PLM to even check the existence
of an ECO, and of course create one if it doesn't yet exists.
- It requires "admin" access levels to Quality to be able to create a
quality point.

For both of those points we need to add a bunch of sudo() accross the
overriden functions, as it makes little sense to add this kind of
permissions to a MRP user just to be able to make a step suggestion.
Also those additional permissions don't modify any existing data, they
only create new data.

Part of task-2985735

Part-of: odoo/enterprise#33880