[FIX] marketing_automation: include the correct file reference in the demo data
Currently in marketing automation, the logo is missing in the template ids
"mass_mailing_offer_free_trial" and "mass_mailing_after_7_days" of the demo
data. This is because there is a wrong file reference in the templates

With a recent refactoring (see commit[1]), it renames the "s_default_image_logo"
with "s_default_image_header_logo" but does not change the file reference in
the template of the demo data in marketing automation.

This commit resolves the above issue by including the correct file reference
in the demo data template.

commit[1] -https://github.com/odoo/odoo/commit/a559fe4
task-3054248

closes odoo/enterprise#33891

Related: odoo/odoo#105124
Signed-off-by: Stéphane Debauche (std) <std@odoo.com>