[IMP] timesheet_grid: use lock dates feature by default
Before this commit, if the user wants to enable the lock
dates feature, he has to enable it in the settings of the
Timesheets app. And when it will be enabled the users will
not can edit a timesheet if the validation date for the
employee linked to the timesheet.

This commit removes the setting in the Timesheets app
settings and directly enable the feature. It means the user
can no longer disabled the feature.

task-3067111

closes odoo/enterprise#34043

Related: odoo/odoo#105905
Related: odoo/upgrade#4040
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>