[FIX] marketing_automation: rewrite hierarchy kanban to OWL specs
Currently, we use a component call in the marketing automation hierarchy
kanbans arch definition to display records in a parent/children relationship.

This is working visually but is not something that the JS framework supports
officially, and that will even break in studio.

The reason for this is that the arch definition of views should be agnostic of
implementation details of its JS class.
This allows the view to work in cases where we don't spawn the JS class
(e.g: studio) and also makes it easier to extend / migrate.

In this commit, we rework this implementation to move all the logic out of the
arch and put it into the view renderer, computing the necessary record "depth"
and display by manually sorting the records and adapting the template.

In addition, and for the same reasons, we remove the various t-on-click calls
to replace them by regular event listeners.

Task-3070918

Part-of: odoo/enterprise#34049
Co-authored-by: Aaron Bohy <aab@odoo.com>