[FIX] account_report: fix reports colspan
In the commit:
https://github.com/odoo/enterprise/commit/72435a6912fc6c62f5a843e8843a1773f1ba5240
A new system for ellipsis in reports has been introduced.
There is a slight issue where if you manually set a colspan
on a line, it would be overriden here anyway. In the journal
report, we rely on manually setting the colspan for grouping
lines (such as the journal line).

By ignoring the calculation if a colspan has already been manually
specified, it will bring back the old behaviour and fix an issue
with the line width in the journal report.

Task id #3063211

closes odoo/enterprise#33974

X-original-commit: f9de978fd06f1b98037dac53ff1085fb8aac8592
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>