[IMP] l10n_us_1099: make list view editable
Before this PR, when we clicked on the list view in the 1099 boxes, a form view would open and you could change the name. Which is basically useless, since we can delete the form view and just edit the list view directly.

Thanks to the change of this PR, it's now possible to edit directly the list view.

closes odoo/enterprise#33921

Related: odoo/upgrade#4031
Signed-off-by: Florian Gilbert (flg) <flg@odoo.com>