[FIX] website_sale_renting: fix unit singular/plural usage
This commit fixes the usages of singular or plural in the label of
rental duration units.

Issue:
- Install website_sale_renting with demo;
- Go on Meeting room product on eCommerce;
- Rent it for several days;

Current behavior:
- Duration unit is displayed in its singular form.

Expected behavior:
- Use duration unit in its correct form.

In the backend, the Odoo standard is to always use the plural form for
units labels. In the frontend/eCommerce, it's expected to use a correct
form. It seems code is written to handle the singular form as an
exception and not use the description_selection method in this
particular case.

Part-of: odoo/enterprise#33931